Properly wait --poll-ldap-interval in poll mode
Instead of a ticker, use a resettable timer for polling the LDAP
server, and reset the timer when a sync happens either because the
config files changed or because the timer went off.

Fix previous behavior, where if a config file change caused a sync in
the middle of the LDAP interval, the next LDAP interval was skipped.